Rename headers that shadow system headers
- addresses modernize-deprecated-headers and hicpp-deprecated-headers
