+++ include/cnl/_impl/fraction/make_fraction.h
@@ -7,8 +7,8 @@
 #if !defined(CNL_IMPL_FRACTION_MAKE_FRACTION_H)
 #define CNL_IMPL_FRACTION_MAKE_FRACTION_H
 
-#include "../../limits.h"
-#include "../assert.h"
+#include "../../numeric_limits.h"
+#include "../cnl_assert.h"
 #include "../type_traits/enable_if.h"
 #include "../type_traits/remove_signedness.h"
 #include "type.h"
