+++ include/cnl/_impl/num_traits/set_digits.h
@@ -10,7 +10,7 @@
 #if !defined(CNL_IMPL_NUM_TRAITS_SET_DIGITS)
 #define CNL_IMPL_NUM_TRAITS_SET_DIGITS
 
-#include "../../limits.h"
+#include "../../numeric_limits.h"
 #include "../type_traits/enable_if.h"
 #include "../type_traits/is_integral.h"
