+++ include/cnl/_impl/to_chars.h
@@ -7,7 +7,7 @@
 #if !defined(CNL_IMPL_TO_CHARS_H)
 #define CNL_IMPL_TO_CHARS_H
 
-#include "assert.h"  // NOLINT(modernize-deprecated-headers,  hicpp-deprecated-headers)
+#include "cnl_assert.h"
 #include "num_traits/digits.h"
 #include "num_traits/rounding.h"
 #include "type_traits/is_signed.h"
