+++ include/cnl/elastic_integer.h
@@ -17,8 +17,8 @@
 #include "_impl/elastic_integer/from_value.h"
 #include "_impl/elastic_integer/generic.h"
 #include "_impl/elastic_integer/is_number.h"
-#include "_impl/elastic_integer/limits.h"
 #include "_impl/elastic_integer/make_elastic_integer.h"
+#include "_impl/elastic_integer/numeric_limits.h"
 #include "_impl/elastic_integer/operators.h"
 #include "_impl/elastic_integer/rep.h"
 #include "_impl/elastic_integer/scale.h"
