+++ test/unit/_impl/wide_integer/definition.cpp
@@ -12,7 +12,7 @@
 #include <cnl/_impl/type_traits/assert_same.h>
 #include <cnl/_impl/type_traits/identical.h>
 #include <cnl/cstdint.h>
-#include <cnl/limits.h>
+#include <cnl/numeric_limits.h>
 #include <cnl/wide_integer.h>
 
 #include <gtest/gtest.h>
