+++ test/unit/limits.cpp
@@ -8,7 +8,7 @@
 /// \brief tests for cnl/limits.h
 
 #include <cnl/cstdint.h>
-#include <cnl/limits.h>
+#include <cnl/numeric_limits.h>
 
 #include <type_traits>
